refactor(core): isolate turn database operations
