+++ crates/openwave-core/src/db/ops/mod.rs
@@ -1,3 +1,4 @@
 pub(in crate::db) mod blob;
 pub(in crate::db) mod conversation;
 pub(in crate::db) mod document;
+pub(in crate::db) mod turn;
